+++ montblanc/arequipa/src/main.rs
@@ -41,7 +41,10 @@ async fn main() -> Result<()> {
 		panic!("Could not create {OUT_FILE}");
 	});
 	let properties = AgentProps { file };
-	let agent = Agent::new(properties).name("arequipa").config(&Config::local()?)?;
+	let agent = Agent::new(properties)
+		.name("arequipa")
+		.prefix("workstation")
+		.config(&Config::local()?)?;
 
 	agent
 		.subscriber()