+++ montblanc/barcelona/src/main.rs
@@ -34,14 +34,17 @@ async fn main() -> Result<()> {
 	init_tracing();
 
 	let properties = AgentProps::default();
-	let agent = Agent::new(properties).name("barcelona").config(&Config::default())?;
+	let agent = Agent::new(properties)
+		.name("barcelona")
+		.prefix("workstation")
+		.config(&Config::default())?;
 
 	agent.publisher().topic("mekong").add()?;
 
 	agent
 		.subscriber()
 		.put_callback(mekong_callback)
-		.topic("mekong")
+		.key_expr("robot/mekong")
 		.add()?;
 
 	agent.start().await?;