+++ montblanc/lyon/src/main.rs
@@ -22,7 +22,10 @@ async fn main() -> Result<()> {
 	init_tracing();
 
 	let properties = AgentProps {};
-	let agent = Agent::new(properties).name("lyon").config(&Config::local()?)?;
+	let agent = Agent::new(properties)
+		.name("lyon")
+		.prefix("robot")
+		.config(&Config::local()?)?;
 
 	agent.publisher().topic("tigris").add()?;