+++ montblanc/osaka/src/main.rs
@@ -54,7 +54,10 @@ async fn main() -> Result<()> {
 	init_tracing();
 
 	let properties = AgentProps::default();
-	let agent = Agent::new(properties).name("osaka").config(&Config::local()?)?;
+	let agent = Agent::new(properties)
+		.name("osaka")
+		.prefix("robot")
+		.config(&Config::local()?)?;
 
 	agent
 		.subscriber()