+++ montblanc/portsmouth/src/main.rs
@@ -16,7 +16,10 @@ async fn main() -> Result<()> {
 	init_tracing();
 
 	let properties = AgentProps {};
-	let agent = Agent::new(properties).name("portsmouth").config(&Config::default())?;
+	let agent = Agent::new(properties)
+		.name("portsmouth")
+		.prefix("robot")
+		.config(&Config::default())?;
 
 	agent.publisher().topic("danube").add()?;