+++ montblanc/taipei/src/main.rs
@@ -24,7 +24,10 @@ async fn main() -> Result<()> {
 	init_tracing();
 
 	let properties = AgentProps {};
-	let agent = Agent::new(properties).name("taipei").config(&Config::local()?)?;
+	let agent = Agent::new(properties)
+		.name("taipei")
+		.prefix("robot")
+		.config(&Config::local()?)?;
 
 	agent.publisher().topic("colorado").add()?;